telegram refactor 1/
move out freqcode from telegram
